rename gooleMapes component, remove commected code
